refactor: lagrange -> polynomial.go
